Find shortest distance using BFS
Write a function, BFS(), to find the shortest distance from vertex v to vertex w, in a directed graph. Vertices ranged from 1 to |V |. The distance is measured by the number of edges. If there is no path from v to w, then -1 is returned. You may assume that the input graph is always valid (no duplicate or any invalid link etc.).