Open addressing hash table
Implement an open addressing hash table by double hashing to perform insertion and deletion. The structure of hash slots is given below and a hash table with 37 hash slots is created in the main function. Both functions’ return value is the number of key comparisons done during insertion and deletion respectively. Inserting a duplicate key will return -1. If the number of key comparisons is more than the table size, it implies that the table is full. For the deletion function, it will return -1 if the deleting key does not exist. The number of key comparisons in deletion cannot be more than the table size. Once the key is deleted, you are not allowed to read it. You only can check its marker if it is a deleted slot.